src/infrastructure/serializers.py
from rest_framework import serializers
from .models import Industry, Technology


class IndustrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Industry
        fields = ["id", "title"]


class TechnologySerializer(serializers.ModelSerializer):
    class Meta:
        model = Technology
        fields = ["id", "title"]
