src/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
celery-redbeat==2.2.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.1
debugpy==1.8.5
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==4.2.9
django-allauth==64.1.0
django-anymail==10.2
django-bootstrap-datepicker-plus==5.0.5
django-debug-toolbar==4.2.0
django-elasticsearch-dsl==8.0
django-extensions==3.2.3
django-import-export==4.1.1
django-mailgun-mime==0.1.7
django-social-auth==0.7.28
django-webpack-loader==3.0.1
djangorestframework==3.15.2
docutils==0.20.1
ecdsa==0.18.0
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
exceptiongroup==1.2.2
Faker==28.0.0
freezegun==1.4.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.6
kombu==5.3.5
Markdown==3.5.2
markdown-urlize==0.2.0
MarkupSafe==2.1.4
msgpack==1.0.7
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-openid==2.2.5
python-social-auth==0.2.12
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.39.2
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
watchgod==0.8.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.6.0
